Fix lin build

+++ src/window/window.cpp
@@ -496,10 +496,10 @@ void window::set_position(const rect &position__, bool redraw_)
 #ifdef _WIN32
     SetWindowPos(context_.hwnd, NULL, position___.left, position___.top, position___.width(), position___.height(), NULL);
 #elif __linux__
-        int32_t values[] = { position___.left,
-            position___.top,
-            position___.width(),
-            position___.height() };
+        uint32_t values[] = { static_cast<uint32_t>(position___.left),
+            static_cast<uint32_t>(position___.top),
+            static_cast<uint32_t>(position___.width()),
+            static_cast<uint32_t>(position___.height()) };
 
         xcb_configure_window(context_.connection,
             context_.wnd,
